+++ src/index.js
@@ -1,11 +1,14 @@
 import {
   tokenize,
   tokenizeWithPunctuation,
-  occurrenceInString,
-  occurrencesInString,
   word,
   punctuation,
   whitespace,
+} from './tokenizers';
+
+import {
+  occurrenceInString,
+  occurrencesInString,
 } from './tokenizers';
 
 import {
